Moved measurement code out of specs and into own module.
