+++ tests/test_monitoring.py
@@ -7,14 +7,15 @@ from pprint import pprint
 from os import environ, getenv
 
 import atlasapi.specs
+import measurements
 from atlasapi.atlas import Atlas
 from json import dumps
 from tests import BaseTests
 import logging
 from time import sleep
 from atlasapi.lib import AtlasUnits, ClusterType
-from atlasapi.specs import AtlasMeasurement, Host, AtlasPeriods, AtlasGranularities, AtlasMeasurementTypes, \
-    AtlasMeasurementValue, ReplicaSetTypes
+from atlasapi.specs import Host, AtlasPeriods, AtlasGranularities, ReplicaSetTypes
+from measurements import AtlasMeasurementTypes, AtlasMeasurementValue, AtlasMeasurement
 from io import BytesIO
 from datetime import datetime, timedelta
 
@@ -64,9 +65,9 @@ class MeasurementTests(BaseTests):
         print(f'For {self.a.Hosts.host_list_with_measurements.__len__()} hosts:')
         for each in self.a.Hosts.host_list_with_measurements[0].measurements:
             print(f'For metric {each.name}')
-            self.assertIsInstance(each.measurement_stats_friendly_bytes, atlasapi.specs.StatisticalValuesFriendly)
+            self.assertIsInstance(each.measurement_stats_friendly_bytes, measurements.StatisticalValuesFriendly)
             print(f'Value is {each.measurement_stats_friendly_bytes.__dict__}')
-            self.assertIsInstance(each.measurement_stats, atlasapi.specs.StatisticalValues)
+            self.assertIsInstance(each.measurement_stats, measurements.StatisticalValues)
 
     test_03_measurement_stats.basic = True
 
@@ -81,9 +82,9 @@ class MeasurementTests(BaseTests):
         print(f'For {self.a.Hosts.host_list_with_measurements.__len__()} hosts:')
         for each in self.a.Hosts.host_list_with_measurements[0].measurements:
             print(f'For metric {each.name}')
-            self.assertIsInstance(each.measurement_stats_friendly_number, atlasapi.specs.StatisticalValuesFriendly)
+            self.assertIsInstance(each.measurement_stats_friendly_number, measurements.StatisticalValuesFriendly)
             print(f'Value is {each.measurement_stats_friendly_number.__dict__}')
-            self.assertIsInstance(each.measurement_stats, atlasapi.specs.StatisticalValues)
+            self.assertIsInstance(each.measurement_stats, measurements.StatisticalValues)
 
     test_04_measurement_stats_objects.basic = True
 
@@ -97,12 +98,12 @@ class MeasurementTests(BaseTests):
         print(f'For {self.a.Hosts.host_list_with_measurements.__len__()} hosts:')
         for each in self.a.Hosts.host_list_with_measurements[0].measurements:
             print(f'For metric {each.name}')
-            self.assertIsInstance(each.measurement_stats_friendly_number, atlasapi.specs.StatisticalValuesFriendly)
+            self.assertIsInstance(each.measurement_stats_friendly_number, measurements.StatisticalValuesFriendly)
             print(f'Value is {each.measurement_stats_friendly_number.__dict__}')
             print(f'For metric {each.name}')
-            self.assertIsInstance(each.measurement_stats_friendly_number, atlasapi.specs.StatisticalValuesFriendly)
+            self.assertIsInstance(each.measurement_stats_friendly_number, measurements.StatisticalValuesFriendly)
             print(f'Value is {each.measurement_stats_friendly_number.__dict__}')
-            self.assertIsInstance(each.measurement_stats, atlasapi.specs.StatisticalValues)
+            self.assertIsInstance(each.measurement_stats, measurements.StatisticalValues)
 
     test_05_measurement_stats_objects_returned.basic = True
 
@@ -116,9 +117,9 @@ class MeasurementTests(BaseTests):
         print(f'For {self.a.Hosts.host_list_with_measurements.__len__()} hosts:')
         for each in self.a.Hosts.host_list_with_measurements[0].measurements:
             print(f'For metric {each.name}')
-            self.assertIsInstance(each.measurement_stats_friendly_bytes, atlasapi.specs.StatisticalValuesFriendly)
+            self.assertIsInstance(each.measurement_stats_friendly_bytes, measurements.StatisticalValuesFriendly)
             print(f'👍Value is {each.measurement_stats_friendly_bytes.__dict__}')
-            self.assertIsInstance(each.measurement_stats, atlasapi.specs.StatisticalValues)
+            self.assertIsInstance(each.measurement_stats, measurements.StatisticalValues)
 
     test_06_measurement_stats_cache_bytes_into.basic = True
 
@@ -132,9 +133,9 @@ class MeasurementTests(BaseTests):
         print(f'For {self.a.Hosts.host_list_with_measurements.__len__()} hosts:')
         for each in self.a.Hosts.host_list_with_measurements[0].measurements:
             print(f'For metric {each.name}')
-            self.assertIsInstance(each.measurement_stats_friendly_bytes, atlasapi.specs.StatisticalValuesFriendly)
+            self.assertIsInstance(each.measurement_stats_friendly_bytes, measurements.StatisticalValuesFriendly)
             print(f'👍Value is {each.measurement_stats_friendly_bytes.__dict__}')
-            self.assertIsInstance(each.measurement_stats, atlasapi.specs.StatisticalValues)
+            self.assertIsInstance(each.measurement_stats, measurements.StatisticalValues)
 
     test_07_measurement_stats_cache_bytes_from.basic = True
 
@@ -148,9 +149,9 @@ class MeasurementTests(BaseTests):
         print(f'For {self.a.Hosts.host_list_with_measurements.__len__()} hosts:')
         for each in self.a.Hosts.host_list_with_measurements[0].measurements:
             print(f'For metric {each.name}')
-            self.assertIsInstance(each.measurement_stats_friendly_bytes, atlasapi.specs.StatisticalValuesFriendly)
+            self.assertIsInstance(each.measurement_stats_friendly_bytes, measurements.StatisticalValuesFriendly)
             print(f'👍Value is {each.measurement_stats_friendly_bytes.__dict__}')
-            self.assertIsInstance(each.measurement_stats, atlasapi.specs.StatisticalValues)
+            self.assertIsInstance(each.measurement_stats, measurements.StatisticalValues)
 
     test_08_measurement_stats_cache_dirty.basic = True
 
@@ -164,9 +165,9 @@ class MeasurementTests(BaseTests):
         print(f'For {self.a.Hosts.host_list_with_measurements.__len__()} hosts:')
         for each in self.a.Hosts.host_list_with_measurements[0].measurements:
             print(f'For metric {each.name}')
-            self.assertIsInstance(each.measurement_stats_friendly_bytes, atlasapi.specs.StatisticalValuesFriendly)
+            self.assertIsInstance(each.measurement_stats_friendly_bytes, measurements.StatisticalValuesFriendly)
             print(f'👍Value is {each.measurement_stats_friendly_bytes.__dict__}')
-            self.assertIsInstance(each.measurement_stats, atlasapi.specs.StatisticalValues)
+            self.assertIsInstance(each.measurement_stats, measurements.StatisticalValues)
 
     test_09_measurement_stats_cache_used.basic = True
 
@@ -180,9 +181,9 @@ class MeasurementTests(BaseTests):
             print(f'For {self.a.Hosts.host_list_with_measurements.__len__()} hosts:')
             for each in self.a.Hosts.host_list_with_measurements[0].measurements:
                 print(f'For metric {each.name}')
-                self.assertIsInstance(each.measurement_stats_friendly_number, atlasapi.specs.StatisticalValuesFriendly)
+                self.assertIsInstance(each.measurement_stats_friendly_number, measurements.StatisticalValuesFriendly)
                 print(f'👍Value is {each.measurement_stats_friendly_number.__dict__}')
-                self.assertIsInstance(each.measurement_stats, atlasapi.specs.StatisticalValues)
+                self.assertIsInstance(each.measurement_stats, measurements.StatisticalValues)
 
     test_10_measurement_stats_tickets.basic = True
 
@@ -195,10 +196,10 @@ class MeasurementTests(BaseTests):
             self.a.Hosts.get_measurement_for_hosts(measurement=measurement, granularity=granularity, period=period)
             print(f'For {self.a.Hosts.host_list_with_measurements.__len__()} hosts:')
             for each in self.a.Hosts.host_list_with_measurements[0].measurements:
-                print(f'For metric {each.name}')
-                self.assertIsInstance(each.measurement_stats_friendly_number, atlasapi.specs.StatisticalValuesFriendly)
+                logger.info(f'For metric {each.name}')
+                self.assertIsInstance(each.measurement_stats_friendly_number, measurements.StatisticalValuesFriendly)
                 print(f'👍Value is {each.measurement_stats_friendly_number.__dict__}')
-                self.assertIsInstance(each.measurement_stats, atlasapi.specs.StatisticalValues)
+                self.assertIsInstance(each.measurement_stats, measurements.StatisticalValues)
 
     test_11_measurement_stats_connections.basic = True
 
@@ -250,9 +251,9 @@ class MeasurementTests(BaseTests):
             for each in self.a.Hosts.host_list_with_measurements[0].measurements:
                 print(f'For metric {each.name}')
                 self.assertIsInstance(each.measurement_stats_friendly_number,
-                                      atlasapi.specs.StatisticalValuesFriendly)
+                                      measurements.StatisticalValuesFriendly)
                 print(f'👍Value is {each.measurement_stats_friendly_number.__dict__}')
-                self.assertIsInstance(each.measurement_stats, atlasapi.specs.StatisticalValues)
+                self.assertIsInstance(each.measurement_stats, measurements.StatisticalValues)
 
     def test_17_return_multiple_metrics(self):
         self.a.Hosts.fill_host_list()
@@ -301,3 +302,12 @@ class MeasurementTests(BaseTests):
                                                            f"returned {type(each_partition)}")
 
     test_21_get_partitions_for_host.basic = True
+
+    def test_22_get_measurements_for_data_partition(self):
+        cluster_name = 'pyAtlasTestCluster'
+        self.a.Hosts.fill_host_list(for_cluster=cluster_name)
+        for each_host in self.a.Hosts.host_list:
+            partition_names = each_host.get_partitions(self.a)
+            for each_partition in partition_names:
+                output = each_host.get_measurements_for_disk(self.a, each_partition)
+                pprint(output)
